Restructure & updates
