+++ .pre-commit-config.yaml
@@ -9,6 +9,21 @@ ci:
   autoupdate_schedule: weekly
 
 repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.4.0
+    hooks:
+      - id: end-of-file-fixer
+      - id: trailing-whitespace
+      # - id: check-json
+      - id: check-yaml
+      - id: check-toml
+      - id: check-docstring-first
+      - id: check-executables-have-shebangs
+      - id: check-case-conflict
+      - id: check-added-large-files
+        args: ["--maxkb=100", "--enforce-all"]
+      - id: detect-private-key
+
   - repo: https://github.com/asottile/pyupgrade
     rev: v3.3.1
     hooks:
@@ -36,3 +51,14 @@ repos:
       - id: flake8
         name: Check PEP8
         args: ["--ignore=E24,W504,F401,E203,W503", "--max-line-length=120",]
+
+  - repo: https://github.com/kynan/nbstripout
+    rev: 0.6.1
+    hooks:
+      - id: nbstripout
+
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.264
+    hooks:
+      - id: ruff
+        args: ["--fix"]
