--- whos_there/__init__.py
@@ -1,3 +1,0 @@
-import pkg_resources
-
-__version__ = pkg_resources.get_distribution("whos_there").version
