--- whos_there/senders/__init__.py
@@ -1,6 +1,0 @@
-from whos_there.senders.debug import DebugSender
-from whos_there.senders.discord import DiscordSender
-from whos_there.senders.email import EmailSender
-from whos_there.senders.slack import SlackSender
-from whos_there.senders.teams import TeamsSender
-from whos_there.senders.telegram import TelegramSender
